Cache getplayerbyid, 403 delete endpoint in prod, remove search endpoint for now

+++ API/Controllers/PlayersController.cs
@@ -14,27 +14,27 @@ namespace API.Controllers
     {
         private readonly EireTrackerContext _context;
         private readonly ICacheService _cache;
+        private readonly IConfiguration _config;
 
-        public PlayersController(EireTrackerContext context, ICacheService cache)
+        public PlayersController(EireTrackerContext context, ICacheService cache, IConfiguration config)
         {
             _context = context;
             _cache = cache;
+            _config = config;
         }
 
         [HttpPost]
         public async Task<ActionResult> AddPlayer([FromBody]Player player)
         {
-            
             _context.Players.Add(player);
-            //await _context.Database.OpenConnectionAsync();
-            //await _context.Database.ExecuteSqlRawAsync(@"SET IDENTITY_INSERT Players ON");
             var isSaved = await _context.SaveChangesAsync() > 0;
-            //await _context.Database.ExecuteSqlRawAsync(@"SET IDENTITY_INSERT Players OFF");
 
             if (isSaved)
+            {
                 return Ok(player);
+            }
 
-            return BadRequest(new ProblemDetails { Title = "Problem saving player" });
+            return BadRequest(new ProblemDetails { Title = "Problem creating player" });
         }
 
         [HttpGet("all")]
@@ -56,9 +56,9 @@ namespace API.Controllers
         [HttpGet("generic/{id}")]
         public async Task<ActionResult<Player>> GetGenericPlayerById(int id)
         {
-            if(!await _context.Players.AnyAsync(x => x.PlayerId == id))
+            if(!await DoesPlayerExist(id))
             {
-                return NoContent();
+                return NotFound();
             }
 
             var player = await _context.Players
@@ -71,52 +71,42 @@ namespace API.Controllers
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<Player>> GetPlayerById(int id)
+        public async Task<ActionResult<PlayerDetailDto>> GetPlayer(int id)
         {
-            if (!await _context.Players.AnyAsync(x => x.PlayerId == id))
+            if (!await _cache.GetOrSetAsync($"player_{id}_exists", () => DoesPlayerExist(id)))
             {
-                return NoContent();
+                return NotFound();
             }
 
-            var player = await _context.Players.Select(x => new PlayerDetailDto
-            {
-                PlayerId = x.PlayerId,
-                Name = x.Name,
-                Club = x.Club,
-                Position = x.Position,
-                DateOfBirth = x.DateOfBirth,
-                OverallStatsDto = x.OverallStats.ConvertToOverallStatsDto(),
-                Performances = x.Performances
-            }).Where(x => x.PlayerId == id)
-              .SingleOrDefaultAsync();
+            var player = await _cache.GetOrSetAsync($"player_{id}_by_id", () => GetPlayerById(id));
 
             return Ok(player);
-        }
+        }  
 
-        [HttpGet("search")]
-        public async Task<IActionResult> SearchPlayers([FromQuery] string name)
-        {
-            if (string.IsNullOrWhiteSpace(name))
-            {
-                return BadRequest(new { error = "Name query parameter is required" });
-            }
+        //[HttpGet("search")]
+        //public async Task<IActionResult> SearchPlayers([FromQuery] string name)
+        //{
+        //    if (string.IsNullOrWhiteSpace(name))
+        //    {
+        //        return BadRequest(new { error = "Name query parameter is required" });
+        //    }
 
-            var players = await _context.Players
-                .Where(p => p.Name.ToLower().Contains(name.ToLower()))
-                .Take(10)
-                .Select(x => new PlayerWithOverallStatsDto
-                {
-                    PlayerId = x.PlayerId,
-                    Name = x.Name,
-                    Club = x.Club,
-                    Position = x.Position,
-                    DateOfBirth = x.DateOfBirth,
-                    OverallStatsDto = x.OverallStats.ConvertToOverallStatsDto()
-                })
-                .ToListAsync();
+        //    var players = await _context.Players
+        //        .Where(p => p.Name.ToLower().Contains(name.ToLower()))
+        //        .Take(10)
+        //        .Select(x => new PlayerWithOverallStatsDto
+        //        {
+        //            PlayerId = x.PlayerId,
+        //            Name = x.Name,
+        //            Club = x.Club,
+        //            Position = x.Position,
+        //            DateOfBirth = x.DateOfBirth,
+        //            OverallStatsDto = x.OverallStats.ConvertToOverallStatsDto()
+        //        })
+        //        .ToListAsync();
 
-            return Ok(players);
-        }
+        //    return Ok(players);
+        //}
 
         [HttpPut]
         public async Task<ActionResult> UpdatePlayerById([FromBody] Player player)
@@ -133,12 +123,20 @@ namespace API.Controllers
         [HttpDelete]
         public async Task<ActionResult> DeleteAll()
         {
+            bool isAllowed = _config.GetValue<bool>("AllowDbDrop");
+            if (!isAllowed)
+            {
+                return Forbid();
+            }
+
             _context.Performances.RemoveRange(_context.Performances);
             _context.OverallStats.RemoveRange(_context.OverallStats);
             _context.Players.RemoveRange(_context.Players);
             await _context.SaveChangesAsync();
             return Ok();
         }
+
+        private async Task<bool> DoesPlayerExist(int id) => await _context.Players.AnyAsync(x => x.PlayerId == id);
 
         private async Task<IEnumerable<PlayerWithOverallStatsDto>> GetOverallStatsOverZeroMinsPlayed()
         {
@@ -154,6 +152,24 @@ namespace API.Controllers
             .ToListAsync();
 
             return players.FindAll(p => p.OverallStatsDto.MinutesPlayed > 0);
+        }
+
+        private async Task<PlayerDetailDto> GetPlayerById(int id)
+        {
+            var player = await _context.Players.Select(x => new PlayerDetailDto
+            {
+                PlayerId = x.PlayerId,
+                Name = x.Name,
+                Club = x.Club,
+                Position = x.Position,
+                DateOfBirth = x.DateOfBirth,
+                OverallStatsDto = x.OverallStats.ConvertToOverallStatsDto(),
+                Performances = x.Performances
+            })
+                .Where(x => x.PlayerId == id)
+                .SingleOrDefaultAsync();
+
+            return player;
         }
     }
 }
